# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reverie API",
    "description": "HTTP API for Reverie, a self-hosted ebook library manager. Generated from the server handlers; do not edit by hand.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/opds/shelves/{shelf_id}": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/shelves/{shelf_id}` — shelf subcatalog navigation root.",
        "description": "# Errors\n- [`AppError::NotFound`] when the shelf is missing or not owned by the\n  caller (existence not leaked).\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "shelf_root",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "description": "Shelf id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS navigation feed linking the shelf's New / Authors / Series subcatalogs",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=navigation": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          },
          "404": {
            "description": "Shelf missing or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    },
    "/opds/shelves/{shelf_id}/authors": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/shelves/{shelf_id}/authors` — the shelf's author index,\ncursor-paginated.",
        "description": "# Errors\n- [`AppError::NotFound`] when the shelf is missing or not owned by the\n  caller (existence not leaked).\n- [`AppError::Validation`] on a malformed cursor.\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "shelf_authors",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "description": "Shelf id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in a previous response's `rel=\"next\"` link;\n`None` returns the first page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS navigation feed of authors with books on the shelf",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=navigation": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          },
          "404": {
            "description": "Shelf missing or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Malformed cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    },
    "/opds/shelves/{shelf_id}/authors/{author_id}": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/shelves/{shelf_id}/authors/{author_id}` — one author's books\non the shelf, cursor-paginated.",
        "description": "# Errors\n- [`AppError::NotFound`] when the shelf is missing or not owned by the\n  caller (existence not leaked).\n- [`AppError::Validation`] on a malformed cursor.\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "shelf_author_books",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "description": "Shelf id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "author_id",
            "in": "path",
            "description": "Author id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in a previous response's `rel=\"next\"` link;\n`None` returns the first page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS acquisition feed of the author's books on the shelf (empty for unknown authors)",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=acquisition": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          },
          "404": {
            "description": "Shelf missing or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Malformed cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    },
    "/opds/shelves/{shelf_id}/new": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/shelves/{shelf_id}/new` — the shelf's newest books,\ncursor-paginated.",
        "description": "# Errors\n- [`AppError::NotFound`] when the shelf is missing or not owned by the\n  caller (existence not leaked).\n- [`AppError::Validation`] on a malformed cursor.\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "shelf_new",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "description": "Shelf id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in a previous response's `rel=\"next\"` link;\n`None` returns the first page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS acquisition feed of the shelf's newest visible books; rel=\"next\" link carries the cursor",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=acquisition": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          },
          "404": {
            "description": "Shelf missing or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Malformed cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    },
...
    "/opds/shelves/{shelf_id}/search": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/shelves/{shelf_id}/search` — full-text search over the\nshelf's books.",
        "description": "# Errors\n- [`AppError::NotFound`] when the shelf is missing or not owned by the\n  caller (existence not leaked).\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "shelf_search",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "description": "Shelf id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search term; an empty (or whitespace-only) query short-circuits\nto an empty feed without hitting the DB.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS acquisition feed of search results scoped to the shelf; empty/whitespace query yields an empty feed",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=acquisition": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          },
          "404": {
            "description": "Shelf missing or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    },
    "/opds/shelves/{shelf_id}/series": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/shelves/{shelf_id}/series` — the shelf's series index,\ncursor-paginated.",
        "description": "# Errors\n- [`AppError::NotFound`] when the shelf is missing or not owned by the\n  caller (existence not leaked).\n- [`AppError::Validation`] on a malformed cursor.\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "shelf_series",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "description": "Shelf id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in a previous response's `rel=\"next\"` link;\n`None` returns the first page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS navigation feed of series with books on the shelf",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=navigation": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          },
          "404": {
            "description": "Shelf missing or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Malformed cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    },
    "/opds/shelves/{shelf_id}/series/{series_id}": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/shelves/{shelf_id}/series/{series_id}` — one series' books\non the shelf, in series order, cursor-paginated.",
        "description": "# Errors\n- [`AppError::NotFound`] when the shelf is missing or not owned by the\n  caller (existence not leaked).\n- [`AppError::Validation`] on a malformed cursor.\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "shelf_series_books",
        "parameters": [
          {
            "name": "shelf_id",
            "in": "path",
            "description": "Shelf id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "series_id",
            "in": "path",
            "description": "Series id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in a previous response's `rel=\"next\"` link;\n`None` returns the first page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS acquisition feed of the series' books on the shelf (empty for unknown series)",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=acquisition": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          },
          "404": {
            "description": "Shelf missing or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Malformed cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 `application/problem+json` error body — the single runtime DTO\nfor the error envelope. The `IntoResponse` impl on\n[`crate::error::AppError`] constructs it for every Problem-Details\nvariant, and operational probes (readiness) build it directly, so the\ndocumented component schema and the bytes on the wire come from the same\nstruct and cannot drift.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "URI reference identifying the specific request, when available.\nOmitted (RFC 9457 §3.1 permits this) outside an HTTP request."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "Stable URI reference identifying the problem type.",
            "example": "https://reverie.example/probs/not-found"
          }
        }
      },
...
    },
    "securitySchemes": {
      "opds_basic": {
        "type": "http",
        "scheme": "basic",
        "description": "HTTP Basic authentication for the OPDS feeds. MUST be used over HTTPS in production — Basic credentials are otherwise exposed in transit (Checkov CKV_OPENAPI_3)."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "id",
        "description": "Session cookie issued by the server's session layer. Set with the Secure attribute when `behind_https` is enabled; MUST be served over HTTPS in production to prevent session hijacking."
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    }
  ],
...
}